% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{spack}
\alias{spack}
\title{Pack small positive integers}
\usage{
spack(x, method = c("doubling", "multiplier4"))
}
\arguments{
\item{x}{A vector of small positive integers.}

\item{method}{String specifying the type of packing to perform.}
}
\value{
A raw vector containing the packed integers.
}
\description{
Pack those small positive integers with a combination of variable-bit and run-length encoding.
}
\examples{
y <- rpois(1000, lambda=0.5) + 1

out <- spack(y)
length(out)

out2 <- spack(y, method="multiplier4")
length(out2)

}
\author{
Aaron Lun
}
